requirements.txt
SQLAlchemy==2.0.40
pandas==2.2.3
Flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.8.5
python-dotenv==1.1.0
Werkzeug==3.1.3
PyMySQL==1.1.1